tflite: python: tflitePythonVivante: Fix build process and remove pre-built libraries
- Fix tim-vx build process which was failing
- Remove pre-built libraries
- Remove unused scripts
- Update base template

+++ tflite/python/tflitePythonVivante/.conf/torizonPackages.ps1
@@ -18,12 +18,16 @@ param()
 
 $TORIZON_ARCH = $args[0]
 
-if ($TORIZON_ARCH -eq "aarch64") { 
-    $TORIZON_ARCH = "arm64" 
+if ($TORIZON_ARCH -eq "aarch64") {
+    $TORIZON_ARCH = "arm64"
 }
 
-if ($TORIZON_ARCH -eq "armv7l") { 
-    $TORIZON_ARCH = "armhf" 
+if ($TORIZON_ARCH -eq "armv7l") {
+    $TORIZON_ARCH = "armhf"
+}
+
+if ($TORIZON_ARCH -eq "x86_64") {
+    $TORIZON_ARCH = "amd64"
 }
 
 # get the files content
@@ -92,14 +96,17 @@ function _ReplaceSection ([string[]]$fileLines, [string]$section) {
 Write-Host "Applying torizonPackages.json ..."
 
 # Dockerfile.debug
-Write-Host "Applying to Dockerfile.debug ..."
-$debugDockerfile = _getFileLines "Dockerfile.debug"
+# The generic project doesn't have a Dockerfile.debug, so check if it exists
+# before applying it
+if (Test-Path -Path "Dockerfile.debug") {
+    Write-Host "Applying to Dockerfile.debug ..."
+    $debugDockerfile = _getFileLines "Dockerfile.debug"
 
-_ReplaceSection $debugDockerfile "torizon_packages_dev" `
-    | Out-File -FilePath "Dockerfile.debug"
+    _ReplaceSection $debugDockerfile "torizon_packages_dev" `
+        | Out-File -FilePath "Dockerfile.debug"
 
-Write-Host -ForegroundColor DarkGreen "✅ Dockerfile.debug"
-
+    Write-Host -ForegroundColor DarkGreen "✅ Dockerfile.debug"
+}
 # Dockerfile.sdk
 # is not all templates that's need the Dockerfile.sdk
 if (Test-Path -Path "Dockerfile.sdk") {
